+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 ## Unreleased
 
 FEATURES:
-* **New Data Source**: d/tfe_organization_workspace_tags is a new datasource to allow reading workspace tags within an organization, by @rhughes1 ([#773](https://github.com/hashicorp/terraform-provider-tfe/pull/773))
+* **New Data Source**: d/tfe_organization_tags is a new data source to allow reading all workspace tags within an organization, by @rhughes1 ([#773](https://github.com/hashicorp/terraform-provider-tfe/pull/773))
 
 ENHANCEMENTS: